refactor: create reusable hooks and components
